refactor(services): extract pure utilities into sub-modules (#69)
- ttsService.js: extract applyVoseoGuard, sanitizeForTTS, splitIntoSentences into ttsHelpers.js
- cuadernoPDF.js: extract PDF template constants into cuadernoPDFTemplates.js
- userProfileService.js: extract PROFILE_QUESTIONS + getApplicableQuestions into profileQuestions.js

Original imports still work via re-exports. No behavior changes.
