+++ src/services/cuadernoPDFTemplates.js
@@ -1,0 +1,20 @@
+export const PAGE_W = 210;
+export const PAGE_H = 297;
+export const MARGIN = 18;
+export const CONTENT_W = PAGE_W - MARGIN * 2;
+
+export const COLOR_EMERALD = [16, 122, 87];
+export const COLOR_SLATE = [71, 85, 105];
+export const COLOR_MUTED = [148, 163, 184];
+export const COLOR_DARK = [15, 23, 42];
+
+export const CUADERNO_VERSION = '1';
+
+export const LOG_TYPE_LABELS = {
+  'log--seeding': 'Siembra',
+  'log--planting': 'Trasplante',
+  'log--harvest': 'Cosecha',
+  'log--input': 'Aplicacion',
+  'log--task': 'Tarea',
+  'log--observation': 'Observacion',
+};
